Make storage of certificates more flexible
This commit changes the way that SSL certificates are stored within
Brass. It provides much greater flexibility, enabling different
certificate users to have different (or multiple) file locations for
different uses. It also simplifies things, by only requiring a single
use to be specified for a certificate attached to a server, rather than
needing to separately define the keys and certificate files.

As it stands, some configuration is needed directly in the database, to
configure the file locations for the different uses of certificates.